requirements.txt
pymodbus
pyyaml 